# improvement: Improve performance of `SemanticTokens`