These are two small changes which greatly improve performance of `SemanticTokensProvider` on big files.

First change is using shared `nodesIterator`. Both lists of tokens and symbols are sorted, so we can advance in them in parallel instead of going through the whole list each time.

Second change is using `SortedSet` for collecting `NodeInfo`.