@jkciesluk 2023-01-10
Using benchmarks from https://github.com/scalameta/metals/pull/4833 the result looks really good
Before changes: 
```
[info] Benchmark                                 (currentHighlightRequest)  Mode  Cnt     Score     Error  Units
[info] SemanticHighlightBench.semanticHighlight                    A.scala    ss   10     1.626 ±   0.411  ms/op
[info] SemanticHighlightBench.semanticHighlight        ClusterDaemon.scala    ss   10   167.772 ±  43.543  ms/op
[info] SemanticHighlightBench.semanticHighlight               Typers.scala    ss   10  2227.971 ± 371.935  ms/op
[info] SemanticHighlightBench.semanticHighlight                Exprs.scala    ss   10    25.480 ±   6.118  ms/op
```

After changes: 
```
[info] Benchmark                                 (currentHighlightRequest)  Mode  Cnt    Score    Error  Units
[info] SemanticHighlightBench.semanticHighlight                    A.scala    ss   10    1.912 ±  0.498  ms/op
[info] SemanticHighlightBench.semanticHighlight        ClusterDaemon.scala    ss   10   47.090 ± 16.405  ms/op
[info] SemanticHighlightBench.semanticHighlight               Typers.scala    ss   10  127.240 ± 45.553  ms/op
[info] SemanticHighlightBench.semanticHighlight                Exprs.scala    ss   10   22.164 ± 17.558  ms/op
```